Merge latest master from upstream

+++ node/lib/repo.js
@@ -1,5 +1,5 @@
 /* global require, console, exports */
-var _       = require('underscore')._,
+var utils   = require('./utils'),
     fs      = require('fs'),
     req     = require('request'),
     path    = require('path'),
@@ -50,7 +50,9 @@ function loadFromCache(url, cachedir, options) {
             log(options).info("Loading from cache: " + url);
             return loadJsonFromFile(path.resolve(cachedir, cache[url].file), options);
         } else {
-            fs.unlink(path.resolve(cachedir, cache[url].date + '.json'));
+            if (fs.existsSync(path.resolve(cachedir, cache[url].date + '.json'))) {
+                fs.unlink(path.resolve(cachedir, cache[url].date + '.json'));
+            }
         }
     }
     var events = new emitter();
@@ -64,7 +66,7 @@ function loadFromCache(url, cachedir, options) {
 }
 
 exports.loadrepository = function(repoUrl, options) {
-    options = _.extend(options, { process : retire.replaceVersion });
+    options = utils.extend(options, { process : retire.replaceVersion });
     if (options.nocache) {
         return loadJson(repoUrl, options);
     }
@@ -72,6 +74,6 @@ exports.loadrepository = function(repoUrl, options) {
 };
 
 exports.loadrepositoryFromFile = function(filepath, options) {
-    options = _.extend(options, { process : retire.replaceVersion });
+    options = utils.extend(options, { process : retire.replaceVersion });
 	return loadJsonFromFile(filepath, options);
 };
